Move get_fieldnames_from_config() to utils
